Fix GetSpellCost
In the case where there is another resource then mana, the table returned by GetSpellPowerCost will have more then 1 index. For the spells that do not cost mana we want to the value from the second index. As far as I can tell at least for shaman there is only 2 resources and no spell costs more then one so is either index 1 or 2.